requirements.txt
anthropic>=0.40.0
python-telegram-bot>=20.0
sqlalchemy>=2.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
